+++ POLLJJAK/src/com/test/controller/MainNoticeController.java
@@ -6,6 +6,7 @@ package com.test.controller;
 
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,41 +30,59 @@ public class MainNoticeController
 	@RequestMapping(value = "/mainnotice.action", method = RequestMethod.GET)
 	public String mainNoticeList(ModelMap model, ArticlePage vo
 			,@RequestParam(value="nowPage" , required=false) String nowPage
-			,@RequestParam(value="cntPerPage", required=false) String cntPerPage)
+			,@RequestParam(value="cntPerPage", required=false) String cntPerPage
+			, HttpServletRequest request)
 	{
 		String result = null;
 		
-		try
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
+		
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
+		
+		if (session.getAttribute("user_code") == null)
 		{
-			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
-			
-			int total = dao.listCount();
-			
-			if (nowPage == null && cntPerPage == null) 
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+		
+			try
 			{
-				nowPage = "1";
-				cntPerPage = "10";
-			} else if (nowPage == null) 
+				IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+				
+				int total = dao.listCount();
+				
+				if (nowPage == null && cntPerPage == null) 
+				{
+					nowPage = "1";
+					cntPerPage = "10";
+				} else if (nowPage == null) 
+				{
+					nowPage = "1";
+				} else if (cntPerPage == null) 
+				{ 
+					cntPerPage = "10";
+				}
+				
+				vo = new ArticlePage(total, Integer.parseInt(nowPage), Integer.parseInt(cntPerPage));
+				
+				MainNoticeDTO dto = new MainNoticeDTO();
+				dto.setStart(Integer.toString(vo.getStart()));
+				dto.setEnd(Integer.toString(vo.getEnd()));
+				model.addAttribute("list", dao.list(dto));
+				model.addAttribute("paging", vo);
+				
+				result = "/Main-Notice.jsp";
+				
+			} catch (Exception e)
 			{
-				nowPage = "1";
-			} else if (cntPerPage == null) 
-			{ 
-				cntPerPage = "10";
+				System.out.println(e.toString());
 			}
-			
-			vo = new ArticlePage(total, Integer.parseInt(nowPage), Integer.parseInt(cntPerPage));
-			
-			MainNoticeDTO dto = new MainNoticeDTO();
-			dto.setStart(Integer.toString(vo.getStart()));
-			dto.setEnd(Integer.toString(vo.getEnd()));
-			model.addAttribute("list", dao.list(dto));
-			model.addAttribute("paging", vo);
-			
-			result = "/Main-Notice.jsp";
-			
-		} catch (Exception e)
-		{
-			System.out.println(e.toString());
 		}
 		return result;
 	}
@@ -72,26 +91,61 @@ public class MainNoticeController
 	public String mainNoticeInsertForm(HttpServletRequest request, Model model)
 	{
 		String result = null;
+		
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
 
-		String a_code = request.getParameter("a_code");
-		model.addAttribute("a_code", a_code);
+		String temp = null; 
 		
-		result = "/Main-Notice-post-insertform.jsp";
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
 		
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+
+			String a_code = request.getParameter("a_code");
+			model.addAttribute("a_code", a_code);
+			
+			result = "/Main-Notice-post-insertform.jsp";
+			
+		}
 		
 		return result;
 	}
 	
 	@RequestMapping(value = "/mainnoticeinsert.action", method = RequestMethod.POST)
-	public String mainNoticeInsert(MainNoticeDTO dto)
+	public String mainNoticeInsert(MainNoticeDTO dto, HttpServletRequest request)
 	{
 		String result = null;
-		IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
-		dao.add(dto);
 		
-		// 리스트가 아니라 해당글상세로 보내야할거같은데..
-		result = "redirect:mainnotice.action";
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
 		
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
+		
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+		
+			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+			dao.add(dto);
+			
+			// 리스트가 아니라 해당글상세로 보내야할거같은데..
+			result = "redirect:mainnotice.action";
+			
+		}
 		
 		return result;
 	}
@@ -101,12 +155,29 @@ public class MainNoticeController
 	{
 		String result = null;
 		
-		IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
-		String notice_code = request.getParameter("notice_code");
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
 		
-		model.addAttribute("noticedetail", dao.noticedetail(notice_code));
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
 		
-		result ="/Main-Notice-post.jsp";
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+
+			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+			String notice_code = request.getParameter("notice_code");
+			
+			model.addAttribute("noticedetail", dao.noticedetail(notice_code));
+			
+			result ="/Main-Notice-post.jsp";
+		}
 		
 		return result;
 	}
@@ -115,41 +186,92 @@ public class MainNoticeController
 	public String mainNoticeUpdateForm(HttpServletRequest request, Model model)
 	{
 		String result = null;
-		IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
 		
-		String notice_code = request.getParameter("notice_code");
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
 		
-		model.addAttribute("notice", dao.noticedetail(notice_code));
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
 		
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
 		
-		result = "/Main-Notice-post-updateform.jsp";
+			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+			
+			String notice_code = request.getParameter("notice_code");
+			
+			model.addAttribute("notice", dao.noticedetail(notice_code));
+			
+			result = "/Main-Notice-post-updateform.jsp";
+		}
 		
 		return result;
 	}
 	
 	@RequestMapping(value="/mainnoticeupdate.action", method=RequestMethod.POST)
-	public String mainNoticeUpdate(MainNoticeDTO dto)
+	public String mainNoticeUpdate(MainNoticeDTO dto, HttpServletRequest request)
 	{
 		String result = null;
-		IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
-		dao.update(dto);
 		
-		result = "redirect:mainnotice.action";
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
+		
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
+		
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+		
+			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+			dao.update(dto);
+			
+			result = "redirect:mainnotice.action";
+		}
 		
 		return result;
 	}
 	
 	
 	@RequestMapping(value="/mainnoticedelete.action", method=RequestMethod.GET)
-	public String mainNoticeRemove(String notice_code)
+	public String mainNoticeRemove(String notice_code, HttpServletRequest request)
 	{
 		String result = null;
 		
-		IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
-		dao.remove(notice_code);
+		// 세션처리 -----------------------------------------
+		HttpSession session = request.getSession();
+
+		String temp = null; 
 		
+		temp = (String) session.getAttribute("user_code");
+		System.out.println(temp);
 		
-		result = "redirect:mainnotice.action";
+		if (session.getAttribute("user_code") == null)
+		{
+			result = "redirect:loginform.action";
+		}
+		else
+		{
+		// ----------------------------------------- 세션처리
+
+			IMainNoticeDAO dao = sqlSession.getMapper(IMainNoticeDAO.class);
+			dao.remove(notice_code);
+			
+			result = "redirect:mainnotice.action";
+		}
 		
 		return result;
 	}
